Check fseek return value and warn if it fails.
It's possible that fseek will fail.  This will at least give an
indication that there is trouble.  It would be better to fail in
gsfAppendIndexFile and return with an error, but this at least a
start.

There are other fseeks without the results checked, but those are
under different CIDs.  They will be handled in future commits.

- Coverity CID 113066
- schwehr/generic-sensor-format#60