Stop wrapping case constants in parens.
The code used both of these styles.

Case 1 - the normal style:

  case FOO_CONST:
    break;

In case 2, this tells the reader that there is something odd going on
behind the scenes that requires FOO_CONST to be protected.  For
example, some sort of macro expansion.

  case (FOO_CONST):
    break;

In this code base, there is never a need to protect those constants,
so I removed all of the parens from case constants.